+++ lib/tracer.dart
@@ -4,3 +4,5 @@ library;
 export 'src/tracer_base.dart';
 
 export 'src/pre/io.dart' if (dart.library.js_interop) 'src/pre/web.dart';
+
+export 'package:stack_trace/src/trace.dart' show Trace;