+++ test/tracer_test.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:tracer/tracer.dart';
 import 'package:test/test.dart';
 
@@ -52,35 +54,40 @@ class TracerTestFatal extends TracerHandler {
 }
 
 void main() {
-  test("Debug", () {
+  test("Debug", () async {
     expect(ranDebug, isFalse);
     final t = Tracer("testing_debug",
         logLevel: TracerLevel.debug, handlers: [TracerTestDebug()]);
     t.debug("This is a test debug message");
+    await Future.delayed(Duration(milliseconds: 100));
     expect(ranDebug, isTrue);
   });
-  test("Info", () {
+  test("Info", () async {
     expect(ranInfo, isFalse);
     final t = Tracer("testing_info", handlers: [TracerTestInfo()]);
     t.info("This is a test info message");
+    await Future.delayed(Duration(milliseconds: 1));
     expect(ranInfo, isTrue);
   });
-  test("Warn", () {
+  test("Warn", () async {
     expect(ranWarn, isFalse);
     final t = Tracer("testing_warn", handlers: [TracerTestWarn()]);
     t.warn("This is a test warn message");
+    await Future.delayed(Duration(milliseconds: 1));
     expect(ranWarn, isTrue);
   });
-  test("Error", () {
+  test("Error", () async {
     expect(ranError, isFalse);
     final t = Tracer("testing_error", handlers: [TracerTestError()]);
     t.error("This is a test error message");
+    await Future.delayed(Duration(milliseconds: 1));
     expect(ranError, isTrue);
   });
-  test("Fatal", () {
+  test("Fatal", () async {
     expect(ranFatal, isFalse);
     final t = Tracer("testing_fatal", handlers: [TracerTestFatal()]);
     t.fatal("This is a test fatal message");
+    await Future.delayed(Duration(milliseconds: 1));
     expect(ranFatal, isTrue);
   });
 }